+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added 
 
 - Data breakpoints
+- Hitcount conditional breakpoints
 - Variables for SystemVerilog
 - Step into
 - Step over